bot/requirements.txt
python-telegram-bot==20.0
solana==0.30.0
solders==0.17.0
